1- Added enginForm template to add an engin. 2- Added validation for engin's attributes.

+++ src/main/java/me/momarija/bioui/service/StorageService.java
@@ -47,6 +47,8 @@ public class StorageService {
 	}
 
 	public void deleteFile(String filename){
+		if (filename.equals("photo1.jpg"))
+			return;
 		try {
 			Files.deleteIfExists(rootLocation.resolve(filename));
 		} catch (IOException e) {
